Move config files back into one file
